+++ mem/vm.h
@@ -9,15 +9,15 @@ void *
 vm_page_reserve(void);
 
 void *
-vm_page_alloc(u64 size);
+vm_page_alloc(size_t size);
 
 void
-vm_page_free(void *page, u64 size);
+vm_page_free(void *page, size_t size);
 
 void *
 vm_page_inquire(void *addr);
 
 void *
-vm_page_extend(void *page, u64 olen, u64 size);
+vm_page_extend(void *page, size_t olen, size_t size);
 
 #endif
